requirements.txt
Django==1.10.4
django-filter==1.0.1
django-geoposition==0.3.0
django-jet==1.0.4
djangorestframework==3.5.3
google-api-python-client==1.5.5
httplib2==0.9.2
Markdown==2.6.7
oauth2client==4.0.0
Pillow==3.4.2
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
six==1.10.0
uritemplate==3.0.0
utils==0.9.0